Draw with the Joystick - Return to (0,0)
This version of the code should implement drawing with the joystick. This should move the motors to the value you specify on the joystick directly. See the scalars in the code to adjust the positions. This means that the joystick resting at center position will return the robot back to position (0,0).